refactor MemberCardController #SD-243

+++ src/main/java/com/example/demo/controller/MemberCardController.java
@@ -3,12 +3,17 @@ package com.example.demo.controller;
 import com.example.demo.model.MemberCard;
 import com.example.demo.service.MemberCardService;
 import lombok.Data;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Map;
+import java.util.UUID;
 
 @Data
 @Validated
@@ -19,18 +24,15 @@ public class MemberCardController {
 
     private final MemberCardService memberCardService;
 
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<MemberCard>> getMemberCards(@RequestParam(name = "page", required = false, defaultValue = "0") Integer page, @RequestParam(name = "size", required = false, defaultValue = "5") Integer size) {
-        return memberCardService.getMemberCards(page, size);
-    }
-
-    @GetMapping(path = "{memberCardUUID}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LinkedHashMap<String, Object>> getMemberCard(@PathVariable("memberCardUUID") UUID memberCardUUID) {
-        return memberCardService.getMemberCard(memberCardUUID);
-    }
-
     @DeleteMapping(value = "/{memberCardUUID}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity deleteMemberCard(@PathVariable("memberCardUUID") UUID memberCardUUID) {
+    public ResponseEntity deleteMemberCard(@PathVariable("memberCardUUID") UUID memberCardUUID, @AuthenticationPrincipal Jwt jwt) {
+
+        String jwtMemberCard = jwt.getClaimAsString("user_memberCardUUID");
+
+        if (!jwtMemberCard.equals(memberCardUUID.toString())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("memberCar UUID mismatch");
+        }
+
         return memberCardService.deleteMemberCard(memberCardUUID);
     }
 
@@ -40,18 +42,39 @@ public class MemberCardController {
     }
 
     @PostMapping(path = "/{memberCardUUID}/borrow", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> postBorrowBooks(@PathVariable("memberCardUUID") UUID memberCardUUID, @RequestBody Map<Object, ArrayList<UUID>> booksArrayJson) {
+    public ResponseEntity<?> postBorrowBooks(@PathVariable("memberCardUUID") UUID memberCardUUID, @RequestBody Map<Object, ArrayList<UUID>> booksArrayJson, @AuthenticationPrincipal Jwt jwt) {
+
+        String jwtMemberCard = jwt.getClaimAsString("user_memberCardUUID");
+
+        if (!jwtMemberCard.equals(memberCardUUID.toString())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("memberCar UUID mismatch");
+        }
+
         return memberCardService.borrowBooks(memberCardUUID, booksArrayJson);
     }
 
     @PostMapping(path = "/{memberCardUUID}/borrow/{borrowUUID}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> returnBorrowBooks(@PathVariable("memberCardUUID") UUID memberCardUUID, @PathVariable("borrowUUID") UUID borrowUUID) {
+    public ResponseEntity<?> returnBorrowBooks(@PathVariable("memberCardUUID") UUID memberCardUUID, @PathVariable("borrowUUID") UUID borrowUUID, @AuthenticationPrincipal Jwt jwt) {
+
+        String jwtMemberCard = jwt.getClaimAsString("user_memberCardUUID");
+
+        if (!jwtMemberCard.equals(memberCardUUID.toString())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("memberCar UUID mismatch");
+        }
+
         return memberCardService.returnBorrowBooks(memberCardUUID, borrowUUID);
     }
 
     @GetMapping(path = "/{memberCardUUID}/history", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> getHistory(@PathVariable("memberCardUUID") UUID memberCardUUID) {
-        return memberCardService.getHistory(memberCardUUID);
+    public ResponseEntity<?> getHistory(@PathVariable("memberCardUUID") UUID memberCardUUID, @AuthenticationPrincipal Jwt jwt, @RequestParam Map<String, ?> allParams) {
+
+//        String jwtMemberCard = jwt.getClaimAsString("user_memberCardUUID");
+
+//        if (!jwtMemberCard.equals(memberCardUUID.toString())) {
+//            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("memberCar UUID mismatch");
+//        }
+
+        return memberCardService.getHistory(memberCardUUID, allParams);
     }
 
 }
